% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\docType{package}
\name{eigenPyR-package}
\alias{eigenPyR}
\alias{eigenPyR-package}
\title{eigenPyR: Eigen library R binding}
\description{
Eigen library R binding
}
\author{
\strong{Maintainer}: Kwan-Young Bak \email{kybak@sungshin.ac.kr} (\href{https://orcid.org/0000-0002-4541-160X}{ORCID}) (Sungshin Women's University) [copyright holder]

Authors:
\itemize{
  \item Kwan-Young Bak \email{kybak@sungshin.ac.kr} (\href{https://orcid.org/0000-0002-4541-160X}{ORCID}) (Sungshin Women's University) [copyright holder]
}

}
